# perf: reduce the overhead of binding function call