--- a/packages/emnapi/src/function.ts
+++ b/packages/emnapi/src/function.ts
@@ -30,7 +30,7 @@ export function napi_get_cb_info (env: napi_env, cbinfo: napi_callback_info, arg
   $CHECK_ENV!(env)
   const envObject = emnapiCtx.envStore.get(env)!
   if (!cbinfo) return envObject.setLastError(napi_status.napi_invalid_arg)
-  const cbinfoValue = emnapiCtx.scopeStore.get(cbinfo)!.getCallbackInfo()!
+  const cbinfoValue = emnapiCtx.scopeStore.get(cbinfo)!.callbackInfo
 
   from64('argc')
   from64('argv')
@@ -176,7 +176,7 @@ export function napi_get_new_target (
 
   from64('result')
 
-  const cbinfoValue = emnapiCtx.scopeStore.get(cbinfo)!.getCallbackInfo()!
+  const cbinfoValue = emnapiCtx.scopeStore.get(cbinfo)!.callbackInfo
   const { thiz, fn } = cbinfoValue
 
   const value = thiz == null || thiz.constructor == null