--- a/packages/runtime/src/HandleScope.ts
+++ b/packages/runtime/src/HandleScope.ts
@@ -1,6 +1,13 @@
 import type { Handle, HandleStore } from './Handle'
 import { External } from './External'
 
+export interface ICallbackInfo {
+  thiz: any
+  data: void_p
+  args: ArrayLike<any>
+  fn: Function
+}
+
 export class HandleScope {
   public handleStore: HandleStore
   public id: number
@@ -8,7 +15,8 @@ export class HandleScope {
   public child: HandleScope | null
   public start: number
   public end: number
-  public _escapeCalled: boolean
+  private _escapeCalled: boolean
+  private _callbackInfo: ICallbackInfo | null
 
   public constructor (handleStore: HandleStore, id: number, parentScope: HandleScope | null, start: number, end = start) {
     this.handleStore = handleStore
@@ -19,6 +27,19 @@ export class HandleScope {
     this.start = start
     this.end = end
     this._escapeCalled = false
+    this._callbackInfo = null
+  }
+
+  public hasCallbackInfo (): boolean {
+    return this._callbackInfo !== null
+  }
+
+  public setCallbackInfo (cbInfo: ICallbackInfo | null) {
+    this._callbackInfo = cbInfo
+  }
+
+  public getCallbackInfo (): ICallbackInfo | null {
+    return this._callbackInfo
   }
 
   public add<V> (value: V): Handle<V> {
@@ -32,6 +53,7 @@ export class HandleScope {
   }
 
   public dispose (): void {
+    if (this._escapeCalled) this._escapeCalled = false
     if (this.start === this.end) return
     this.handleStore.erase(this.start, this.end)
   }