About 1.5x faster in average

macOS 14.4.1 23E224 arm64 Apple M2 Pro 32GB

Node.js v22.13.1

Before

```
binding: function () {}
                 emnapi #emptyFunction x 40,783,722 ops/sec ±0.83% (98 runs sampled)

binding: function (obj) { return obj }
                 emnapi #returnParam x 21,951,118 ops/sec ±0.54% (94 runs sampled)
```

After

```
binding: function () {}
                 emnapi #emptyFunction x 63,958,599 ops/sec ±1.10% (94 runs sampled)

binding: function (obj) { return obj }
                 emnapi #returnParam x 29,147,277 ops/sec ±0.65% (93 runs sampled)
```